moves FakeAction to common location
